refactor(playback): extract clear_lyrics_display to sibling file (file-size cap)
Follow-up to 85488de. The Phase A.1 changes pushed
crates/sp-server/src/playback/mod.rs from 993 → 1014 lines, over the
1000-line file-size cap enforced by CI's File Size Check.

Extracts clear_lyrics_display into the existing sibling-include pattern
already used by position_update.rs / recovery.rs. Identical behavior;
the function is just relocated. mod.rs is now 977 lines.
